phase-2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend with SQLModel ORM and SQL database support"
requires-python = ">=3.11"
dependencies = [
   "fastapi>=0.128.0",
   "psycopg2-binary>=2.9.11",
   "psycopg[binary]>=3.3.2",
   "pyjwt>=2.10.1",
   "python-dotenv>=1.2.1",
   "sqlmodel>=0.0.31",
   # Database drivers (install based on your database)
   # "pymysql",         # MySQL
   # "pyodbc",          # SQL Server
   "uvicorn>=0.40.0",
]

